convert AnalyzeTalkTopicsJob to talk.agents.analyse_topics
