+++ app/models/talk.rb
@@ -108,11 +108,6 @@ class Talk < ApplicationRecord
     language.present? ? Language.find(language)&.alpha2 : Language::DEFAULT
   end
 
-  # TODO convert to performs
-  def analyze_talk_topics!
-    AnalyzeTalkTopicsJob.perform_now(self)
-  end
-
   # search
   # meilisearch do
   #   attribute :title
