--- test/jobs/analyze_talk_topics_job_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class AnalyzeTalkTopicsJobTest < ActiveJob::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
